Adds ability to leave a review.
Creates review model and basic flow for creating a review and viewing
existing reviews.

The idea of the flow is a user comes to the page, leaves a review and
looks at other reviews.

Since the idea is a person would leave only one review and we don't want
reviews to be edited after the fact I removed actions such as updating
and destroying.